WIP: Subscription and delete data is now working
* Lot of rough edges

+++ main/remoteservices/src/EBox/RemoteServices/Subscription.pm
@@ -46,13 +46,13 @@ use Cwd;
 use Error qw(:try);
 use File::Slurp;
 use File::Temp;
+use JSON::XS;
 use HTML::Mason;
 use Net::Ping;
 
 # Constants
 use constant {
     SERV_CONF_DIR => 'remoteservices',
-    SERV_SUBDIR => 'remoteservices/subscription',
     SERV_CONF_FILE => 'remoteservices.conf',
     PROF_PKG       => 'zentyal-cloud-prof',
     SEC_UPD_PKG    => 'zentyal-security-updates',
@@ -85,6 +85,11 @@ sub new
 
     $self->{user} = $params{user};
     $self->{password} = $params{password};
+
+    # Set the REST client
+    $self->{restClient} = new EBox::RemoteServices::RESTClient(
+        credentials => { username => $params{user},
+                         password => $params{password} });
 
     bless $self, $class;
     return $self;
@@ -140,12 +145,14 @@ sub soapCall
                        );
 }
 
-# Method: subscribeEBox
+# Method: subscribeServer
 #
 #      Given a name trying to subscribe a server for that user using
-#      that name. If it is already registered, the process will
-#      fail. If the process works nicely, a bundle is got which is
-#      used to set the parameters to connect to the Zentyal Cloud
+#      that name. If it is already registered, the process will reset
+#      the server password. If the process works nicely, a set of
+#      basic parameters are stored. In order to get the full
+#      functionality, it is required to get the bundle in a separated
+#      process to set the parameters to connect to the Zentyal Cloud
 #      infrastructure including the required certificates.
 #
 # Parameters:
@@ -161,12 +168,7 @@ sub soapCall
 #
 #      hash ref - containing the following keys and values:
 #
-#        ca - the CA certificate path
-#        cert - the certificate path for this eBox
-#        key - the private key path for this eBox
-#
-#        confFile - the configuration file to be used to connect to
-#        the infrastructure
+#        confFile - the configuration file to be used to connect to#        the infrastructure
 #
 #        new - Boolean indicating if the subscription was done or just
 #        the file getting
@@ -180,11 +182,11 @@ sub soapCall
 #      <EBox::Exceptions::MissingArgument> - thrown if the compulsory
 #      argument is missing
 #
-sub subscribeEBox
+sub subscribeServer
 {
-    my ($self, $cn, $option) = @_;
+    my ($self, $name, $option) = @_;
 
-    $cn or throw EBox::Exceptions::MissingArgument('cn');
+    $name or throw EBox::Exceptions::MissingArgument('name');
 
     # Ensure firewall rules are opened
     $self->_openHTTPSConnection();
@@ -192,10 +194,10 @@ sub subscribeEBox
     # Check the WS is reachable
     $self->_checkWSConnectivity();
 
-    $option = undef if ( $option eq 'reload' );
+    $option = undef if ( defined($option) and ( $option eq 'reload' ));
 
     if (not $option) {
-        my $availables = $self->_getAvailable($cn);
+        my $availables = $self->_getAvailable($name);
 
         my $checker = new EBox::RemoteServices::Subscription::Check();
         # Check the available editions are suitable for this server
@@ -221,61 +223,70 @@ sub subscribeEBox
         }
     }
 
-    my $vpnSettings;
-    try {
-        $vpnSettings = $self->soapCall('vpnSettings',
-                                       option => $option);
-    } catch EBox::Exceptions::DataNotFound with { };
-    unless ( defined($vpnSettings) ) {
-        throw EBox::Exceptions::External(
-            __x(
-                'Cannot retrieve VPN settings needed for subscription. Check your {openurl}{brand} profile{closeurl} to check your VPN server settings.',
-                brand    => 'Zentyal Cloud',
-                openurl  => q{<a href='https://cloud.zentyal.com/services/profile/'>},
-                closeurl => q{</a>}
-               )
-           );
-    }
+    # my $vpnSettings;
+    # try {
+    #     $vpnSettings = $self->soapCall('vpnSettings',
+    #                                    option => $option);
+    # } catch EBox::Exceptions::DataNotFound with { };
+    # unless ( defined($vpnSettings) ) {
+    #     throw EBox::Exceptions::External(
+    #         __x(
+    #             'Cannot retrieve VPN settings needed for subscription. Check your {openurl}{brand} profile{closeurl} to check your VPN server settings.',
+    #             brand    => 'Zentyal Cloud',
+    #             openurl  => q{<a href='https://cloud.zentyal.com/services/profile/'>},
+    #             closeurl => q{</a>}
+    #            )
+    #        );
+    # }
 
-    # Check the selected VPN server is reachable
-    $self->_checkVPNConnectivity(
-                                 $vpnSettings->{ipAddr},
-                                 $vpnSettings->{protocol},
-                                 $vpnSettings->{port},
-                                );
+    # # Check the selected VPN server is reachable
+    # $self->_checkVPNConnectivity(
+    #                              $vpnSettings->{ipAddr},
+    #                              $vpnSettings->{protocol},
+    #                              $vpnSettings->{port},
+    #                             );
+
+    # my $conf;
+    my $response = $self->{restClient}->POST("/v1/servers/",
+                                             { 'name' => $name, 'option' => $option} );
+    my $serverInfoRaw = $response->as_string();
+    my $serverInfo = $response->data();
+
+    # Write this info to a file only readable by Zentyal
+    $self->_writeCredentials($name, $serverInfoRaw);
+
+    return { 'new' => $serverInfo->{created} };
+
+    # my $new = 0;
+    # my $rs = EBox::Global->modInstance('remoteservices');
+    # try {
+    #     $ = $self->soapCall('subscribeEBox',
+    #                                      canonicalName => $name,
+    #                                      rsVersion     => $rs->version(),
+    #                                      option        => $option);
+    #     $new = 1;
+    # } catch EBox::Exceptions::DataExists with {
+    #     $bundleRawData = $self->soapCall('eBoxBundle',
+    #                                      canonicalName => $name,
+    #                                      rsVersion     => $rs->version(),
+    #                                      option        => $option);
+    #     $new = 0;
+    # };
+
+    # my $params = $self->extractBundle($name, $bundleRawData);
+
+    # my $confKeys = EBox::Config::configKeysFromFile($params->{confFile});
+    # $self->_openVPNConnection(
+    #     $confKeys->{vpnServer},
+    #     $confKeys->{vpnPort},
+    #     $confKeys->{vpnProtocol},
+    #    );
 
 
-    my $bundleRawData;
-    my $new = 0;
-    my $rs = EBox::Global->modInstance('remoteservices');
-    try {
-        $bundleRawData = $self->soapCall('subscribeEBox',
-                                         canonicalName => $cn,
-                                         rsVersion     => $rs->version(),
-                                         option        => $option);
-        $new = 1;
-    } catch EBox::Exceptions::DataExists with {
-        $bundleRawData = $self->soapCall('eBoxBundle',
-                                         canonicalName => $cn,
-                                         rsVersion     => $rs->version(),
-                                         option        => $option);
-        $new = 0;
-    };
+    # $self->executeBundle($params, $confKeys);
 
-    my $params = $self->extractBundle($cn, $bundleRawData);
-
-    my $confKeys = EBox::Config::configKeysFromFile($params->{confFile});
-    $self->_openVPNConnection(
-        $confKeys->{vpnServer},
-        $confKeys->{vpnPort},
-        $confKeys->{vpnProtocol},
-       );
-
-
-    $self->executeBundle($params, $confKeys);
-
-    $params->{new} = $new;
-    return $params;
+    # $params->{new} = $new;
+    # return $params;
 }
 
 # Method: serversList
@@ -355,20 +366,8 @@ sub extractBundle
     my @files = $tar->list_files();
     my $cwd = Cwd::getcwd();
 
-    my $dirPath = EBox::Config::conf() . SERV_CONF_DIR;
-    unless ( -d $dirPath ) {
-        mkdir($dirPath);
-    }
-    $dirPath = EBox::Config::conf() . SERV_SUBDIR;
-    unless (chdir($dirPath)) {
-        mkdir($dirPath);
-        chdir($dirPath);
-    }
-    $dirPath .= "/$cn";
-    unless (chdir($dirPath)) {
-        mkdir($dirPath);
-        chdir($dirPath);
-    }
+    my $dirPath = $class->_createSubscriptionDir($cn);
+    chdir($dirPath);
 
     my ($confFile, $keyFile, $certFile, $qaSources, $qaGpg, $qaPreferences, $installCloudProf);
     my @scripts;
@@ -458,6 +457,10 @@ sub executeBundle
     $self->_setDDNSConf();
     $self->_installCloudProf($params, $confKeys);
     $self->_executeBundleScripts($params, $confKeys);
+
+    # Set to have the bundle
+    my $rs = EBox::Global->getInstance()->modInstance('remoteservices');
+    $rs->st_set_bool('has_bundle', 1);
 }
 
 # Method: deleteData
@@ -487,7 +490,7 @@ sub deleteData
 
     $cn or throw EBox::Exceptions::MissingArgument('cn');
 
-    my $dirPath = EBox::Config::conf() . SERV_SUBDIR . "/$cn";
+    my $dirPath = $self->_subscriptionDirPath($cn);
 
     unless ( -d $dirPath ) {
         return 0;
@@ -501,35 +504,38 @@ sub deleteData
     closedir($dir);
     rmdir($dirPath);
 
-    # Remove QA updates configuration
-    $self->_removeQAUpdates();
-    # Remove DDNS autoconfiguration
-    $self->_removeDDNSConf();
-    # Remove alert autoconfiguration
-    # FIXME: Do by alertAutoconfiguration script?
-    my $events = EBox::Global->modInstance('events');
-    $events->unset('alert_autoconfiguration');
-    # TODO: Remove zentyal-cloud-prof package
+    my $rs = EBox::Global->modInstance('remoteservices');
+    if ( $rs->hasBundle() ) {
+        # Remove QA updates configuration
+        $self->_removeQAUpdates();
+        # Remove DDNS autoconfiguration
+        $self->_removeDDNSConf();
+        # Remove alert autoconfiguration
+        # FIXME: Do by alertAutoconfiguration script?
+        my $events = EBox::Global->modInstance('events');
+        $events->unset('alert_autoconfiguration');
+        # TODO: Remove zentyal-cloud-prof package
 
-    # Remove jobs
-    my $cronPrefix = EBox::RemoteServices::Configuration::CronJobPrefix();
-    $dirPath = EBox::Config::conf() . SERV_CONF_DIR . '/jobs';
-    if ( -d $dirPath ) {
-        opendir($dir, $dirPath);
-        while(my $filePath = readdir($dir)) {
-            next unless -d "$dirPath/$filePath";
-            if ($filePath =~ m/^[0-9]+$/g or $filePath =~ m/^$cronPrefix/g) {
-                EBox::Sudo::command("rm -rf '$dirPath/$filePath'");
-            } elsif ( $filePath =~ m/incoming|outcoming/g ) {
-                # Remove any left incoming/outcoming job
-                EBox::Sudo::command("rm -f '$dirPath/$filePath/*'");
+        # Remove jobs
+        my $cronPrefix = EBox::RemoteServices::Configuration::CronJobPrefix();
+        $dirPath = EBox::Config::conf() . SERV_CONF_DIR . '/jobs';
+        if ( -d $dirPath ) {
+            opendir($dir, $dirPath);
+            while(my $filePath = readdir($dir)) {
+                next unless -d "$dirPath/$filePath";
+                if ($filePath =~ m/^[0-9]+$/g or $filePath =~ m/^$cronPrefix/g) {
+                    EBox::Sudo::command("rm -rf '$dirPath/$filePath'");
+                } elsif ( $filePath =~ m/incoming|outcoming/g ) {
+                    # Remove any left incoming/outcoming job
+                    EBox::Sudo::command("rm -f '$dirPath/$filePath/*'");
+                }
             }
+            closedir($dir);
         }
-        closedir($dir);
     }
 
     # Remove subscription levels and disaster recovery if any
-    my $rs = EBox::Global->modInstance('remoteservices');
+    $rs->st_unset('has_bundle');
     $rs->st_delete_dir('subscription');
     $rs->st_delete_dir('disaster_recovery');
 
@@ -1065,13 +1071,39 @@ sub _getAvailable
 {
     my ($self, $server) = @_;
 
-    my $client = new EBox::RemoteServices::RESTClient(
-        credentials => { username => $self->{user},
-                         password => $self->{password}});
-
-    my $response = $client->GET("/v1/bundle/available/$server/");
+    my $response = $self->{restClient}->GET("/v1/bundle/available/$server/");
     return $response->data();
+}
 
+# Write received credentials into the proper directory and set proper permissions
+sub _writeCredentials
+{
+    my ($self, $name, $serverInfoRaw) = @_;
+
+    $self->_createSubscriptionDir($name);
+
+    my $credentialsFilePath = $self->_credentialsFilePath($name);
+
+    File::Slurp::write_file($credentialsFilePath, $serverInfoRaw);
+    chmod(0600, $credentialsFilePath);
+}
+
+# Create the subscription directory structure based on the given name
+# Return the newly created directory
+sub _createSubscriptionDir
+{
+    my ($self, $name) = @_;
+
+    my $dirPath = $self->_subscriptionDirPath($name);
+    unless ( -d $dirPath ) {
+        my @dirs;
+        my @path = split(/\//, $dirPath);
+        foreach my $dir (@path) {
+            push(@dirs, $dir);
+            mkdir(join('/', @dirs));
+        }
+    }
+    return $dirPath;
 }
